resolve requested change
